requirements.txt
inflection==0.3.1
peewee==2.8.3
py==1.4.31
pycodestyle==2.3.1
pytest==3.0.1
python-bitcoinrpc==1.0
simplejson==3.8.2
colorama==0.3.9
termcolor==1.1.0
pynput==1.3.7
